requirements.txt
# pip freeze
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
pycodestyle==2.10.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.1
urllib3==1.26.14
yarl==1.8.2
# 